refactor: 위젯 위치 변경 (features -> widgets)
