+++ src/features/post/index.ts
@@ -1,2 +1,1 @@
 export * from "./model";
-export * from "./ui";
