+++ src/features/user/model/index.ts
@@ -1,2 +1,3 @@
 export * from "./userQueryKey";
 export * from "./useUserQuery";
+export * from "./useUserStore";
