+++ src/widgets/comment/CommentEditDialog.tsx
@@ -1,10 +1,8 @@
-import { Dialog, DialogContent, DialogTitle } from "@radix-ui/react-dialog";
-import { useCommentStore } from "../model/useCommentStore";
-import { Button, DialogHeader, Textarea } from "@shared/ui";
+import { Dialog, DialogContent, DialogTitle, Button, DialogHeader, Textarea } from "@shared/ui";
 import { Comment, updateComment } from "@entities/comment/model";
 import { useQueryClient } from "@tanstack/react-query";
 import { usePostStore } from "@features/post";
-import { commentQueryKey } from "../model";
+import { commentQueryKey, useCommentStore } from "@features/comment";
 
 const CommentEditDialog = () => {
   const { showEditCommentDialog, setShowEditCommentDialog, selectedComment, setSelectedComment } = useCommentStore();
