+++ src/widgets/index.ts
@@ -1,0 +1,4 @@
+export * from "./comment";
+export * from "./post";
+export * from "./ui";
+export * from "./user";
