+++ src/widgets/post/index.ts
